More sugar through lazy helper object
This patch adds the ability for helpers to be used in a subject, by allowing
for lazy query results. Helpers that execute in a subject execute at rspec
file parse time, so we must delays the subsequent '#run' request on a node
after startup. These new helper objects will delay there #run until the
data is accessed to avoid this ordering issue.

This patch also removes the early setup block, as this no longer works with
these helpers. This was deprecated earlier, so this removal means this change
is now a major API changing one.

Signed-off-by: Ken Barber <ken@bob.sh>